split out concurrency submodule
