+++ smart_open/s3.py
@@ -8,12 +8,10 @@
 """Implements file-like objects for reading and writing from/to S3."""
 
 import io
-import contextlib
 import functools
 import logging
 import time
 import urllib.parse
-import warnings
 
 import boto
 import boto3
@@ -21,32 +19,12 @@ import botocore.client
 import botocore.exceptions
 
 import smart_open.bytebuffer
+import smart_open.concurrency
 import smart_open.utils
 
 from smart_open import constants
 
 logger = logging.getLogger(__name__)
-
-# AWS Lambda environments do not support multiprocessing.Queue or multiprocessing.Pool.
-# However they do support Threads and therefore concurrent.futures's ThreadPoolExecutor.
-# We use this flag to allow python 2 backward compatibility, where concurrent.futures doesn't exist.
-_CONCURRENT_FUTURES = False
-try:
-    import concurrent.futures
-    _CONCURRENT_FUTURES = True
-except ImportError:
-    warnings.warn("concurrent.futures could not be imported and won't be used")
-
-# Multiprocessing is unavailable in App Engine (and possibly other sandboxes).
-# The only method currently relying on it is iter_bucket, which is instructed
-# whether to use it by the MULTIPROCESSING flag.
-_MULTIPROCESSING = False
-try:
-    import multiprocessing.pool
-    _MULTIPROCESSING = True
-except ImportError:
-    warnings.warn("multiprocessing could not be imported and won't be used")
-
 
 DEFAULT_MIN_PART_SIZE = 50 * 1024**2
 """Default minimum part size for S3 multipart uploads"""
@@ -1042,7 +1020,7 @@ def iter_bucket(
         retries=retries,
         **session_kwargs)
 
-    with _create_process_pool(processes=workers) as pool:
+    with smart_open.concurrency.create_pool(processes=workers) as pool:
         result_iterator = pool.imap_unordered(download_key, key_iterator)
         for key_no, (key, content) in enumerate(result_iterator):
             if True or key_no % 1000 == 0:
@@ -1125,41 +1103,3 @@ def _download_fileobj(bucket, key_name):
     buf = io.BytesIO()
     bucket.download_fileobj(key_name, buf)
     return buf.getvalue()
-
-
-class DummyPool(object):
-    """A class that mimics multiprocessing.pool.Pool for our purposes."""
-    def imap_unordered(self, function, items):
-        return map(function, items)
-
-    def terminate(self):
-        pass
-
-
-class ConcurrentFuturesPool(object):
-    """A class that mimics multiprocessing.pool.Pool but uses concurrent futures instead of processes."""
-    def __init__(self, max_workers):
-        self.executor = concurrent.futures.ThreadPoolExecutor(max_workers)
-
-    def imap_unordered(self, function, items):
-        futures = [self.executor.submit(function, item) for item in items]
-        for future in concurrent.futures.as_completed(futures):
-            yield future.result()
-
-    def terminate(self):
-        self.executor.shutdown(wait=True)
-
-
-@contextlib.contextmanager
-def _create_process_pool(processes=1):
-    if _MULTIPROCESSING and processes:
-        logger.info("creating multiprocessing pool with %i workers", processes)
-        pool = multiprocessing.pool.Pool(processes=processes)
-    elif _CONCURRENT_FUTURES and processes:
-        logger.info("creating concurrent futures pool with %i workers", processes)
-        pool = ConcurrentFuturesPool(max_workers=processes)
-    else:
-        logger.info("creating dummy pool")
-        pool = DummyPool()
-    yield pool
-    pool.terminate()
